Extract date section into a separate file
